refactor: convert Hds::Text::Body to typescript
